+++ src/extension.ts
@@ -1,8 +1,9 @@
 import {commands, ExtensionContext, languages, Uri, workspace} from 'vscode';
 import App from './app';
-import Resolver from './getters-setters/resolver';
-import Builder from './fabric/builder';
-import Documenter from './phpdoc/documenter';
+import Construct from './feature/constructor/construct';
+import Builder from './feature/fabric/builder';
+import Resolver from './feature/getters-setters/resolver';
+import Documenter from './feature/phpdoc/documenter';
 import {
     CMD_GENERATE_ABSTRACT_CLASS,
     CMD_GENERATE_CLASS,
@@ -11,13 +12,14 @@ import {
     CMD_GENERATE_FINAL_CLASS,
     CMD_GENERATE_INTERFACE,
     CMD_GENERATE_PHPDOC,
+    CMD_GENERATE_PHPDOC_MASTER,
     CMD_GENERATE_TRAIT,
     CMD_INSERT_GETTER,
     CMD_INSERT_GETTER_SETTER,
-    CMD_INSERT_GETTER_SETTER_WIZARD,
-    CMD_INSERT_GETTER_WIZARD,
+    CMD_INSERT_GETTER_SETTER_MASTER,
+    CMD_INSERT_GETTER_MASTER,
     CMD_INSERT_SETTER,
-    CMD_INSERT_SETTER_WIZARD,
+    CMD_INSERT_SETTER_MASTER,
     F_ABSTRACT_CLASS,
     F_CLASS,
     F_ENUM,
@@ -27,7 +29,6 @@ import {
     R_GETTER,
     R_SETTER,
 } from './constants';
-import Construct from './constructor/construct';
 
 export async function activate(context: ExtensionContext) {
     /* getters-setters */
@@ -46,17 +47,17 @@ export async function activate(context: ExtensionContext) {
         const resolver = new Resolver([position]);
         resolver.render([R_GETTER, R_SETTER]);
     }));
-    context.subscriptions.push(commands.registerCommand(CMD_INSERT_GETTER_WIZARD, async () => {
+    context.subscriptions.push(commands.registerCommand(CMD_INSERT_GETTER_MASTER, async () => {
         const positions = await Resolver.selectProperties('Select properties to generate getters');
         const resolver = new Resolver(positions);
         resolver.render([R_GETTER]);
     }));
-    context.subscriptions.push(commands.registerCommand(CMD_INSERT_SETTER_WIZARD, async () => {
+    context.subscriptions.push(commands.registerCommand(CMD_INSERT_SETTER_MASTER, async () => {
         const positions = await Resolver.selectProperties('Select properties to generate setters');
         const resolver = new Resolver(positions);
         resolver.render([R_SETTER]);
     }));
-    context.subscriptions.push(commands.registerCommand(CMD_INSERT_GETTER_SETTER_WIZARD, async () => {
+    context.subscriptions.push(commands.registerCommand(CMD_INSERT_GETTER_SETTER_MASTER, async () => {
         const positions = await Resolver.selectProperties('Select properties to generate getters and setters');
         const resolver = new Resolver(positions);
         resolver.render([R_GETTER, R_SETTER]);
@@ -91,7 +92,12 @@ export async function activate(context: ExtensionContext) {
     /* phpdoc */
     context.subscriptions.push(commands.registerCommand(CMD_GENERATE_PHPDOC, () => {
         const position = App.instance.editor.selection.active;
-        const documenter = new Documenter(position);
+        const documenter = new Documenter([position]);
+        documenter.render();
+    }));
+    context.subscriptions.push(commands.registerCommand(CMD_GENERATE_PHPDOC_MASTER, async () => {
+        const positions = await Documenter.selectBlocks('Select blocks to generate phpdocs');
+        const documenter = new Documenter(positions);
         documenter.render();
     }));
 
@@ -103,38 +109,28 @@ export async function activate(context: ExtensionContext) {
     }));
 
     /* symfony */
+    const watcher = workspace.createFileSystemWatcher('**/config/services.{yml,yaml}');
+    watcher.onDidChange((uri) => App.instance.symfony.updateServices(uri));
+    watcher.onDidCreate((uri) => App.instance.symfony.updateServices(uri));
+    context.subscriptions.push(watcher);
+    App.instance.providers.forEach((p) => {
+        context.subscriptions.push(languages.registerCodeLensProvider(p.selector, p.provider));
+    });
+    await App.instance.symfony.updateServices(await symfonyServicesYamlUri());
+}
+
+async function symfonyServicesYamlUri(): Promise<Uri|null> {
     const candidates = ['config/services.yaml', 'config/services.yml', 'app/config/services.yaml'];
-    let servicesYamlUri = null;
     for (const candidate of candidates) {
         try {
             const uri = Uri.joinPath(workspace.workspaceFolders![0].uri, candidate);
             await workspace.fs.stat(uri);
-            servicesYamlUri = uri;
-            break;
-        } catch (e) {
-            servicesYamlUri = null;
-        }
-    }
-    if (servicesYamlUri !== null) {
-        await App.instance.updateSymfonyServices(servicesYamlUri);
+
+            return uri;
+        } catch (e) {}
     }
 
-    const watcher = workspace.createFileSystemWatcher('**/config/services.{yml,yaml}');
-    watcher.onDidChange((uri) => App.instance.updateSymfonyServices(uri));
-    watcher.onDidCreate((uri) => App.instance.updateSymfonyServices(uri));
-    context.subscriptions.push(watcher);
-    context.subscriptions.push(languages.registerCodeLensProvider(
-        {language: 'php'},
-        App.instance.symfonyServicesProvider,
-    ));
-    context.subscriptions.push(languages.registerCodeLensProvider(
-        {language: 'php'},
-        App.instance.symfonyTemplatesProvider,
-    ));
-    context.subscriptions.push(languages.registerCodeLensProvider(
-        {language: 'yaml', pattern: '**/config/services.{yml,yaml}'},
-        App.instance.symfonyServicesYamlProvider,
-    ));
+    return null;
 }
 
 export function deactivate() {}